Move interp1d into init so Sphinx automodule works
